+++ packages/veui/src/components/Carousel.vue
@@ -1,44 +1,50 @@
 <template>
 <div class="veui-carousel" :ui="ui">
   <div class="veui-carousel-viewport" @mouseenter="handleEnter" @mouseleave="handleLeave">
-  <transition-group name="veui-carousel-item" class="veui-carousel-items" tag="ol">
-    <li v-for="(item, i) in datasource" v-show="localIndex === i" :key="i"
-      :class="{
-        'veui-carousel-item': true,
-        'veui-carousel-item-current': localIndex === i
-      }"
-      :style="{
-        'background-image': `url(${item.src})`
+    <transition-group name="veui-carousel-item" class="veui-carousel-items" tag="ol">
+      <li
+        v-for="(item, i) in datasource"
+        v-show="localIndex === i"
+        ref="item"
+        :key="i"
+        :class="{
+          'veui-carousel-item': true,
+          'veui-carousel-item-current': localIndex === i
+        }"
+        :style="{
+          'background-image': `url(${item.src})`
+        }"
+        tabindex="0">
+        <slot v-bind="item" :index="i">
+          <img class="veui-sr-only" :src="item.src" :alt="item.alt">
+        </slot>
+      </li>
+    </transition-group>
+    <div v-if="indicator === 'number'" class="veui-carousel-indicator-numbers">{{ localIndex + 1 }}<span class="veui-carousel-indicator-numbers-separator"></span>{{ count }}</div>
+    <nav v-else-if="indicator !== 'none'" :class="{
+        [`veui-carousel-indicator-${indicator}s`]: true
       }">
-      <slot v-bind="item" :index="i">
-        <img :src="item.src" :alt="item.alt">
-      </slot>
-    </li>
-  </transition-group>
-  <div v-if="indicator === 'number'" class="veui-carousel-indicator-numbers">{{ localIndex + 1 }}<span class="veui-carousel-indicator-numbers-separator"></span>{{ count }}</div>
-  <nav v-else-if="indicator !== 'none'" :class="{
-      [`veui-carousel-indicator-${indicator}s`]: true
-    }">
-    <button type="button" v-for="(item, i) in datasource" :key="i"
-      :class="{
-        'veui-carousel-indicator-item': true,
-        'veui-carousel-indicator-item-current': localIndex === i
-      }"
-      @click="select(i, 'click')"
-      @mouseenter="select(i, 'hover')"
-    >{{ item.label || `第 ${i + 1} 页` }}</button>
-  </nav>
-  <button type="button" class="veui-carousel-control veui-carousel-control-prev"
-    @click="step(-1)"
-    :disabled="!wrap && localIndex === 0">
-    <veui-icon :name="icons.prev"/>
-  </button>
-  <button type="button" class="veui-carousel-control veui-carousel-control-next"
-    @click="step(1)"
-    :disabled="!wrap && localIndex === count - 1">
-    <veui-icon :name="icons.next"/>
-  </button>
+      <button type="button" v-for="(item, i) in datasource" :key="i"
+        :class="{
+          'veui-carousel-indicator-item': true,
+          'veui-carousel-indicator-item-current': localIndex === i
+        }"
+        @click="select(i, 'click')"
+        @mouseenter="select(i, 'hover')"
+      >{{ item.label || `第 ${i + 1} 页` }}</button>
+    </nav>
+    <button type="button" class="veui-carousel-control veui-carousel-control-prev"
+      @click="step(-1)"
+      :disabled="!wrap && localIndex === 0">
+      <veui-icon :name="icons.prev" label="上一页"/>
+    </button>
+    <button type="button" class="veui-carousel-control veui-carousel-control-next"
+      @click="step(1)"
+      :disabled="!wrap && localIndex === count - 1">
+      <veui-icon :name="icons.next" label="下一页"/>
+    </button>
   </div>
+  <div class="veui-sr-only" aria-live="polite" aria-atomic="true">当前是第 {{ localIndex + 1 }} 页，共 {{ datasource.length }} 页</div>
 </div>
 </template>
 
@@ -122,6 +128,12 @@ export default {
     select (index, event) {
       if (event !== this.switchTrigger) {
         return
+      }
+
+      if (event === 'click') {
+        setTimeout(() => {
+          this.$refs.item[this.localIndex].focus()
+        }, 0)
       }
 
       this.localIndex = index