+++ packages/veui/src/components/Radio.vue
@@ -20,7 +20,7 @@ export default {
   inheritAttrs: false,
   mixins: [ui, input],
   model: {
-    prop: 'checked',
+    prop: 'model',
     event: 'change'
   },
   props: {
@@ -28,11 +28,12 @@ export default {
       type: null,
       default: true
     },
-    checked: null
+    checked: Boolean,
+    model: null
   },
   data () {
     return {
-      localChecked: this.checked
+      localChecked: this.model === this.value ? true : this.checked
     }
   },
   computed: {
@@ -46,13 +47,31 @@ export default {
     }
   },
   watch: {
-    checked (value) {
-      this.localChecked = value
+    checked: {
+      handler (val) {
+        this.localChecked = val
+      },
+      immediate: true
     },
-    localChecked (value) {
-      if (this.checked !== value) {
-        this.$emit('change', value)
-      }
+    localChecked: {
+      handler (val) {
+        if (this.checked !== val) {
+          this.$emit('update:checked', val)
+        }
+
+        if (val) {
+          this.$emit('change', this.value)
+        }
+      },
+      immediate: true
+    },
+    model: {
+      handler (val) {
+        if (val != null) {
+          this.localChecked = val === null ? false : this.value === val
+        }
+      },
+      immediate: true
     }
   }
 }