+++ packages/veui/src/components/RadioGroup.vue
@@ -5,19 +5,20 @@
   role="radiogroup"
   :aria-readonly="String(realReadonly)"
   :aria-disabled="String(realDisabled)">
-  <radio
+  <veui-radio
     :ui="ui"
     :name="localName"
     v-for="(item, index) in items"
     :key="index"
     :value="item.value"
+    :model="value"
     :disabled="item.disabled || realDisabled || realReadonly"
     :checked="item.value === value"
-    @change="checked => handleChange(item.value, checked)"
+    @change="model => handleChange(item.value, model)"
     :aria-posinset="index + 1"
     :aria-setsize="items.length">
     <slot v-bind="item">{{ item.label }}</slot>
-  </radio>
+  </veui-radio>
 </div>
 </template>
 
@@ -30,7 +31,7 @@ import { uniqueId } from 'lodash'
 export default {
   name: 'veui-radio-group',
   components: {
-    'radio': Radio
+    'veui-radio': Radio
   },
   mixins: [ui, input],
   model: {
@@ -46,7 +47,8 @@ export default {
     }
   },
   methods: {
-    handleChange (value, checked) {
+    handleChange (value, model) {
+      let checked = value === model
       if (checked) {
         this.$emit('change', value)
       }