+++ packages/veui/src/components/Table/Column.vue
@@ -2,7 +2,7 @@
 import { uniqueId, pick } from 'lodash'
 import table from '../../mixins/table'
 import { getIndexOfType } from '../../utils/context'
-import '../../config/uiTypes'
+import '../../common/uiTypes'
 
 export default {
   name: 'veui-table-column',
@@ -29,7 +29,9 @@ export default {
       id: uniqueId('veui-table-column-')
     }
   },
-  render () {},
+  render () {
+    return null
+  },
   created () {
     let index = getIndexOfType(this, 'table-column')